Feat: change index and reservation pages
