+++ pages/index.js
@@ -1,25 +1,9 @@
-import Head from "next/head";
-import Image from "next/image";
-import styles from "../styles/Home.module.css";
+import { reservationMachine } from "../utils/state-machine";
+import { useMachine } from "@xstate/react";
+import ReservationController from "../components/ReservationController";
 
-export default function Home() {
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>DiverList</title>
-        <meta name="description" content="Top diving places in the world" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
+const Reservation = () => {
+  return <ReservationController />;
+};
 
-      <main className={styles.main}>
-        <h1 className={styles.title}>Best Diving places in the World!</h1>
-        <p className={styles.description}>
-          Get started by editing{" "}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
-
-        <div className={styles.grid}></div>
-      </main>
-    </div>
-  );
-}
+export default Reservation;
